Refactor SimulationTagData interface so that it resides within interfaces folder
- Creating the UML package diagram was what prompted this change. Based on making this diagram, I realised that it made more sense to move this interface to the interfaces folder.
